Crate spec_runner

spec-runner is the ruby/spec runner for Artichoke.

spec-runner is a wrapper around MSpec and ruby/spec that works with the Artichoke virtual filesystem. spec-runner runs the specs declared in a manifest file.

Spec Manifest

spec-runner is invoked with a YAML manifest that specifies which specs to run. The manifest can run whole suites, like all of the StringScanner specs, or specific specs, like the Array#drop spec. The manifest supports marking specs as skipped.

core:
  - suite: array
    specs:
      - any
      - append
      - array
  - suite: comparable
  - suite: string
    specs:
      - scan
library:
  - suite: stringscanner
  - suite: uri
    skip:
      - parse

Usage

$ cargo run -q -p spec-runner -- --help
spec-runner 0.1.0
ruby/spec runner for Artichoke.

USAGE:
    spec-runner <config>

FLAGS:
    -h, --help       Prints help information
    -V, --version    Prints version information

ARGS:
    <config>    Path to YAML config file

Modules

model
Models for reading spec manifests.

mspec
Embedded MSpec framework.

rubyspec
Embedded copy of ruby/spec suites.

Structs

Opt
CLI specification for spec-runner.

Functions

is_require_path
Determine if an embedded ruby/spec should be tested.

main
Main entrypoint.

try_main
Result-returning entrypoint.